feat(api): containerize platform and expand church workflows

+++ src/modules/churchs/services/DeleteChurchService.ts
@@ -1,15 +1,10 @@
 import { IChurch } from "../../../entities/IChurch";
 import NoExistError from "../../../shared/errors/NoExistError";
 import { IChurchRepository } from "../repositories/IChurchRepository";
-import { ILocationRepository } from "../repositories/ILocationRepository";
 
 export default class DeleteChurchService {
-  constructor(
-    private churchRepository: IChurchRepository,
-    private locationRepository: ILocationRepository
-  ) {
+  constructor(private churchRepository: IChurchRepository) {
     this.churchRepository = churchRepository;
-    this.locationRepository = locationRepository;
   }
 
   public async execute(id_church: number) : Promise<IChurch | undefined>{
@@ -17,16 +12,14 @@ export default class DeleteChurchService {
 
     if (!church)
       throw new NoExistError("church")
+
+    if (church.type === "HEADQUARTER")
+      throw new Error("Headquarter can not be deleted")
     
     const isChurchDeleted = await this.churchRepository.delete(id_church);
 
     if (!isChurchDeleted)
       throw new Error("Church doesn't deleted")
-
-    const isLocationDeleted = await this.locationRepository.delete(church.id_location);
-
-    if (!isLocationDeleted)
-      throw new Error("Location doesn't deleted")
 
     return church;
   }
